Initial Ramen commit
Here we see Ramen working by showing a question when you view `View 2`.

We use `Math.random()` when creating the ID and email, because Ramen
will only show a question one time to a given user by default.

By using `Math.random()` the question will appear every time because you
are effectively a different User.

In the next commit, we will show how to tie in the ramenSettings
object to the currently logged in user.